requirements.txt
Flask
psycopg2-binary
selenium
undetected-chromedriver
python-dotenv
APScheduler